ITB-211: Refactored paging - GET call fixed!
