+++ src/main/java/com/inthebytes/restaurantmanager/dao/RestaurantDao.java
@@ -1,5 +1,7 @@
-package com.inthebytes.restaurantmanager.repository;
+package com.inthebytes.restaurantmanager.dao;
 
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -9,4 +11,5 @@ import com.inthebytes.restaurantmanager.entity.Restaurant;
 public interface RestaurantDao extends JpaRepository<Restaurant, String> {
 	Restaurant findByRestaurantId(String id);
 	Restaurant findByName(String name);
+	Page<Restaurant> findAll(Pageable pageable);
 }
