+++ src/main/java/com/inthebytes/restaurantmanager/dao/UserDao.java
@@ -1,4 +1,4 @@
-package com.inthebytes.restaurantmanager.repository;
+package com.inthebytes.restaurantmanager.dao;
 
 import com.inthebytes.restaurantmanager.entity.User;
